+++ Classes/Crawler.php
@@ -1,0 +1,54 @@
+<?php
+
+declare(strict_types=1);
+
+namespace AOE\Crawler;
+
+/*
+ * (c) 2020 AOE GmbH <dev@aoe.com>
+ *
+ * This file is part of the TYPO3 Crawler Extension.
+ *
+ * It is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU General Public License, either version 2
+ * of the License, or any later version.
+ *
+ * For the full copyright and license information, please read the
+ * LICENSE.txt file that was distributed with this source code.
+ *
+ * The TYPO3 project - inspiring people to share!
+ */
+
+use TYPO3\CMS\Core\Core\Environment;
+use TYPO3\CMS\Core\SingletonInterface;
+use TYPO3\CMS\Core\Utility\GeneralUtility;
+
+final class Crawler implements SingletonInterface
+{
+    /**
+     * @var string
+     */
+    private $processFilename;
+
+    public function __construct(string $processFilename = null)
+    {
+        $this->processFilename = $processFilename ?: Environment::getVarPath() . '/lock/tx_crawler.proc';
+        $this->setDisabled(false);
+    }
+
+    public function setDisabled($disabled = true): void
+    {
+        if ($disabled) {
+            GeneralUtility::writeFile($this->processFilename, '');
+        } else {
+            if (is_file($this->processFilename)) {
+                unlink($this->processFilename);
+            }
+        }
+    }
+
+    public function isDisabled(): bool
+    {
+        return is_file($this->processFilename);
+    }
+}